simulators/portal: refactor sims into 1 sim multi suite format
