+++ simulators/portal/beacon/Dockerfile
@@ -1,8 +1,10 @@
-FROM rust:1.71.1 AS builder
+FROM rust:1.75.0 AS builder
 
 # create a new empty shell project
-RUN USER=root cargo new --bin rpc-compat
-WORKDIR /rpc-compat
+RUN USER=root cargo new --bin beacon
+WORKDIR /beacon
+
+RUN apt-get update && apt-get install clang -y
 
 # copy over manifests and source to build image
 COPY Cargo.toml ./Cargo.toml
@@ -17,8 +19,8 @@ FROM ubuntu:22.04
 RUN apt update && apt install wget -y
 
 # copy build artifacts from build stage
-COPY --from=builder /rpc-compat/target/release/rpc-compat .
+COPY --from=builder /beacon/target/release/beacon .
 
 ENV RUST_LOG=debug
 
-ENTRYPOINT ["./rpc-compat"]
+ENTRYPOINT ["./beacon"]
