+++ simulators/portal/beacon/src/suites/constants.rs
@@ -1,2 +1,5 @@
 pub const HIVE_PORTAL_NETWORKS_SELECTED: &str = "HIVE_PORTAL_NETWORKS_SELECTED";
 pub const BEACON_STRING: &str = "beacon";
+
+// trin-bridge constants
+pub const TRIN_BRIDGE_CLIENT_TYPE: &str = "trin-bridge";
