+++ simulators/portal/beacon/src/suites/mod.rs
@@ -1,0 +1,2 @@
+pub mod constants;
+pub mod rpc_compat;
