--- simulators/portal/history/portal-interop/src/constants.rs
@@ -1,1 +1,0 @@
-pub const TEST_DATA_FILE_PATH: &str = "./test-data/test_data_collection_of_forks_blocks.yaml";
