--- simulators/portal/history/portal-mesh/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "portal-mesh"
-version = "0.1.0"
-authors = ["Kolby ML (Moroz Liebl) <kolbydml@gmail.com>"]
-edition = "2021"
-
-[dependencies]
-ethportal-api = { git = "https://github.com/ethereum/trin", rev = "2a32224e3c2b0b80bc37c1b692c33016371f197a" }
-hivesim = { git = "https://github.com/ethereum/portal-hive", rev = "8ff1e3d3c941dd00d56dacd777a5dfb71edf402f" }
-itertools = "0.10.5"
-serde_json = "1.0.87"
-tokio = { version = "1", features = ["full"] }
-tracing = "0.1.37"
-tracing-subscriber = "0.3.16"
