+++ simulators/portal/history/src/suites/mesh.rs
@@ -1,10 +1,12 @@
+use crate::suites::constants::TRIN_BRIDGE_CLIENT_TYPE;
 use ethportal_api::jsonrpsee::core::__reexports::serde_json;
 use ethportal_api::types::distance::{Metric, XorMetric};
 use ethportal_api::types::portal::ContentInfo;
 use ethportal_api::{
     Discv5ApiClient, HistoryContentKey, HistoryContentValue, HistoryNetworkApiClient,
 };
-use hivesim::{dyn_async, Client, NClientTestSpec, Simulation, Suite, Test, TestSpec};
+use hivesim::types::ClientDefinition;
+use hivesim::{dyn_async, Client, NClientTestSpec, Test};
 use itertools::Itertools;
 use serde_json::json;
 use std::collections::HashMap;
@@ -17,46 +19,12 @@ const HEADER_WITH_PROOF_VALUE: &str = "0x080000002d020000f90222a02c58e3212c08517
 // private key hive environment variable
 const PRIVATE_KEY_ENVIRONMENT_VARIABLE: &str = "HIVE_CLIENT_PRIVATE_KEY";
 
-#[tokio::main]
-async fn main() {
-    tracing_subscriber::fmt::init();
-
-    let mut suite = Suite {
-        name: "portal-mesh".to_string(),
-        description: "The portal mesh test suite runs a set of scenarios to test 3 clients"
-            .to_string(),
-        tests: vec![],
-    };
-
-    suite.add(TestSpec {
-        name: "Portal Network mesh".to_string(),
-        description: "".to_string(),
-        always_run: false,
-        run: test_portal_scenarios,
-        client: None,
-    });
-
-    let sim = Simulation::new();
-    run_suite(sim, suite).await;
-}
-
-async fn run_suite(host: Simulation, suite: Suite) {
-    let name = suite.clone().name;
-    let description = suite.clone().description;
-
-    let suite_id = host.start_suite(name, description, "".to_string()).await;
-
-    for test in &suite.tests {
-        test.run_test(host.clone(), suite_id, suite.clone()).await;
-    }
-
-    host.end_suite(suite_id).await;
-}
-
 dyn_async! {
-   async fn test_portal_scenarios<'a> (test: &'a mut Test, _client: Option<Client>) {
+   pub async fn test_portal_scenarios<'a> (test: &'a mut Test, _client: Option<Client>) {
         // Get all available portal clients
         let clients = test.sim.client_types().await;
+        // todo: remove this once we implement role in hivesim-rs
+        let clients: Vec<ClientDefinition> = clients.into_iter().filter(|client| client.name != *TRIN_BRIDGE_CLIENT_TYPE).collect();
 
         let private_key_1 = "fc34e57cc83ed45aae140152fd84e2c21d1f4d46e19452e13acc7ee90daa5bac".to_string();
         let private_key_2 = "e5add57dc4c9ef382509e61ce106ec86f60eb73bbfe326b00f54bf8e1819ba11".to_string();
