--- simulators/portal/history/trin-bridge/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "trin-bridge"
-version = "0.1.0"
-authors = ["Kolby ML (Moroz Liebl) <kolbydml@gmail.com>"]
-edition = "2021"
-
-[dependencies]
-ethportal-api = { git = "https://github.com/ethereum/trin", rev = "b530dc3d40d51c21c800089eacb2852ebd8c4d45" }
-portal-spec-test-utils-rs = { git = "https://github.com/ethereum/portal-spec-tests", rev = "d1e996d0d4dc2136b3cd38d9e25cdc3a6b74dcd9" }
-hivesim = { git = "https://github.com/ethereum/portal-hive", rev = "8ff1e3d3c941dd00d56dacd777a5dfb71edf402f" }
-itertools = "0.10.5"
-serde_yaml = "0.9"
-tokio = { version = "1", features = ["full"] }
-tracing = "0.1.37"
-tracing-subscriber = "0.3.16"
